Initial commit.
This repo will contain various code snippets used for various purposes.
Some of these are troubleshooting, others are proof-of-concepts, etc.
There is currently a single file that has now been included in the repo;
this is a prototype for a factory class that I have in my Foresight
project.  In that case, the code has been crashing and I cannot
understand the reason for that.  This was an attempt to test out the
approach being used there, without the baggage of existing code and
possibly cumbersome existing class hierarchy. It is noteworthy that this
test worked spectacularly!